requirements.txt
cleverminer
pandas
openpyxl
psycopg2
python-dotenv